requirements.txt
django
mysqlclient
pillow
django-crispy-forms
crispy-bootstrap5
djangorestframework